added support for metadata in transform functions
Added a new configuration environment variable called
COUCHIMPORT_META. If present, it's parsed as a JSON string
and stored in the couchimport config object. It is then
passed as a second argument to transform functions. It shouldn't
affect existing transform functions. For example, if the current
parse filename is in a bash variable 'f', the file name can be passed
to the transform function like this:

  export COUCHIMPORT_META="{\"file_name\":\"$f\"}"