Updated couchexport examples
In couchexport examples, I added the URL, which is necessary for locating the database. I also split this into one CSV example and one TSV example and called out that you need to specify a comma as the delimiter for the CSV but not for TSV (where tab is the default).